serverless/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
torchvision==0.16.0+cpu
Pillow==10.0.0
numpy==1.24.4
psutil==5.9.5  # Added for CPU metrics
boto3==1.34.14
aws-xray-sdk==2.12.1
flask==3.0.0
flask-cors==4.0.0
rich==13.7.0